move gitHubWorker and refactor some logic into an issue model file
