+++ client/components/TicketList.js
@@ -2,7 +2,7 @@ const React = require('react');
 const TicketSearch = require('./TicketSearch');
 const TicketEntry = require('./ticketEntry');
 const $ = require('jquery');
-const GH = require('../../gitHubWorker');
+const Issues = require('../js/issues');
 
 class TicketList extends React.Component {
   
@@ -10,7 +10,8 @@ class TicketList extends React.Component {
     super(props);
     
     this.state = {
-      tickets: []
+      allTickets: [],
+      ticketsToRender: []
     };
     
     this.getIssues = this.getIssues.bind(this);
@@ -21,9 +22,10 @@ class TicketList extends React.Component {
   getIssues(searchTerm, language){
     //Fetch issues;
     var self = this;
-    GH.getIssues(function(data) {
+    Issues.getIssues(function(data) {
       self.setState({
-        tickets: data
+        allTickets: data,
+        ticketsToRender: data.slice(0,199)
       });
     }, console.log, searchTerm, language);
   }
@@ -38,7 +40,7 @@ class TicketList extends React.Component {
       <TicketSearch searchHandler={this.getIssues} />
       <h4>Open Issues</h4>
       <div className="main-ticket-view">
-          {this.state.tickets.map ((ticket, index) => 
+          {this.state.ticketsToRender.map ((ticket, index) => 
             <TicketEntry data={ticket} key={index} />
           )}
       </div>
